Exceptions Treatments for Insert,Update and Delete. (Just for training and example)

+++ src/main/java/com/enzoccs/SpringBootWebProject/services/UserService.java
@@ -4,10 +4,15 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.enzoccs.SpringBootWebProject.entities.User;
 import com.enzoccs.SpringBootWebProject.repositories.UserRepository;
+import com.enzoccs.SpringBootWebProject.services.exceptions.DatabaseException;
+import com.enzoccs.SpringBootWebProject.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class UserService {
@@ -22,7 +27,7 @@ public class UserService {
 	
 	public User findById(Integer id) {
 		Optional<User> u = repository.findById(id);
-		return u.get();
+		return u.orElseThrow(()-> new ResourceNotFoundException(id));
 	}
 	
 	public User insertUser(User u) {
@@ -30,12 +35,32 @@ public class UserService {
 	}
 	
 	public void deleteUser(Integer id) {
-		repository.deleteById(id);
+		try {
+			
+			if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+			
+			repository.deleteById(id);
+		
+		}catch(DataIntegrityViolationException e){
+			throw new DatabaseException(e.getMessage());
+		}catch(ResourceNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		}
+		
 	}
 	
 	public User updateUser(User u ,Integer id) {
-		User userToUpdate = repository.getReferenceById(id);
-		updateUser(userToUpdate,u);
+		User userToUpdate = null;
+		
+		try {
+			userToUpdate = repository.getReferenceById(id);
+			updateUser(userToUpdate,u);
+
+		}catch(EntityNotFoundException e){
+			throw new ResourceNotFoundException(id);
+		}
+		
+		
 		return repository.save(userToUpdate);
 	}
